+++ BackgroundButtonIcon/src/java/example/MainPanel.java
@@ -127,7 +127,7 @@ class ArrowToggleButtonBarCellIcon implements Icon {
         int h = c.getHeight() - 1;
         int h2 = (int) (h * .5 + .5);
         int w2 = TH;
-        Path2D p = new Path2D.Float();
+        Path2D p = new Path2D.Double();
         p.moveTo(0,      0);
         p.lineTo(w - w2, 0);
         p.lineTo(w,      h2);
