+++ RoundedTextField/src/java/example/MainPanel.java
@@ -83,11 +83,11 @@ class RoundedCornerBorder extends AbstractBorder {
         Graphics2D g2 = (Graphics2D) g.create();
         g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
         int r = height - 1;
-        RoundRectangle2D round = new RoundRectangle2D.Float(x, y, width - 1, height - 1, r, r);
+        RoundRectangle2D round = new RoundRectangle2D.Double(x, y, width - 1, height - 1, r, r);
         Container parent = c.getParent();
         if (Objects.nonNull(parent)) {
             g2.setPaint(parent.getBackground());
-            Area corner = new Area(new Rectangle2D.Float(x, y, width, height));
+            Area corner = new Area(new Rectangle2D.Double(x, y, width, height));
             corner.subtract(new Area(round));
             g2.fill(corner);
         }
